backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
wagtail==6.4
psycopg2==2.9.10
djangorestframework_simplejwt==5.4.0
django-cors-headers==4.7.0
gunicorn==21.2.0
boto3==1.38.8
django-storages==1.14.6
django-filter~=24.3
python-decouple==3.8
python-dotenv==1.1.1
django-phonenumber-field[phonenumbers]==7.3.0
